Initialize Delivery API project, deploy database with Docker Compose, and establish connection
Set up the initial structure for the web API project and introduced a Docker Compose file to deploy the database. Additionally, included the latest copy of the web API in the Docker Compose file, enabling users to deploy both the web API and the database together if desired. The commit also encompasses the integration of the database into the project using a connection string.